LPS-54791 Categories navigation: Standardize locations for internal setting classes
