+++ modules/apps/asset/asset-categories-navigation-web/src/com/liferay/asset/categories/navigation/web/settings/internal/AssetCategoriesNavigationPortletInstanceSettingsIdMapping.java
@@ -12,10 +12,11 @@
  * details.
  */
 
-package com.liferay.asset.categories.navigation.web.configuration.internal;
+package com.liferay.asset.categories.navigation.web.settings.internal;
 
 import com.liferay.asset.categories.navigation.web.configuration.AssetCategoriesNavigationPortletInstanceConfiguration;
-import com.liferay.portal.kernel.settings.definition.ConfigurationBeanDeclaration;
+import com.liferay.asset.categories.navigation.web.constants.AssetCategoriesNavigationPortletKeys;
+import com.liferay.portal.kernel.settings.definition.SettingsIdMapping;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -23,13 +24,17 @@ import org.osgi.service.component.annotations.Component;
  * @author Juergen Kappler
  */
 @Component
-public class
-	AssetCategoriesNavigationPortletInstanceConfigurationBeanDeclaration
-		implements ConfigurationBeanDeclaration {
+public class AssetCategoriesNavigationPortletInstanceSettingsIdMapping
+	implements SettingsIdMapping {
 
 	@Override
-	public Class getConfigurationBeanClass() {
+	public Class<?> getConfigurationBeanClass() {
 		return AssetCategoriesNavigationPortletInstanceConfiguration.class;
+	}
+
+	@Override
+	public String getSettingsId() {
+		return AssetCategoriesNavigationPortletKeys.ASSET_CATEGORIES_NAVIGATION;
 	}
 
 }
